chore(webapp): move Adhoc/Continuous pages to /pages
