+++ webapp/javascript/pages/Adhoc.module.scss
@@ -1,5 +1,3 @@
-// This file is called "Adhoc" since it should apply to all pages
-
 // All panels should have the same height
 // So that when tabs are changed there's no content shift
 .tabPanel {
