+++ webapp/javascript/pages/ContinuousComparisonView.jsx
@@ -5,19 +5,19 @@ import 'react-dom';
 import { bindActionCreators } from 'redux';
 import Box from '@ui/Box';
 import { FlamegraphRenderer } from '@pyroscope/flamegraph';
-import TimelineChartWrapper from './TimelineChartWrapper';
-import Header from './Header';
-import Footer from './Footer';
+import TimelineChartWrapper from '../components/TimelineChartWrapper';
+import Header from '../components/Header';
+import Footer from '../components/Footer';
 import { buildRenderURL } from '../util/updateRequests';
 import {
   fetchNames,
   fetchComparisonAppData,
   fetchTimeline,
 } from '../redux/actions';
-import InstructionText from './InstructionText';
-import ExportData from './ExportData';
-import useExportToFlamegraphDotCom from './exportToFlamegraphDotCom.hook';
-import styles from './ComparisonApp.module.css';
+import InstructionText from '../components/InstructionText';
+import ExportData from '../components/ExportData';
+import useExportToFlamegraphDotCom from '../components/exportToFlamegraphDotCom.hook';
+import styles from './ContinuousComparison.module.css';
 
 // See docs here: https://github.com/flot/flot/blob/master/API.md
